Fix port numbers, fix drive doing integer division, add tray fourbar

+++ src/drive.cpp
@@ -17,7 +17,7 @@ Drive::Drive(okapi::MotorGroup leftMotors, okapi::MotorGroup rightMotors) {
 }
 
 void Drive::run(pros::Controller& joystick) {
-    double forward = joystick.get_analog(pros::E_CONTROLLER_ANALOG_LEFT_Y) / 127;
-    double turn = joystick.get_analog(pros::E_CONTROLLER_ANALOG_RIGHT_X) / 127;
+    double forward = (double)(joystick.get_analog(pros::E_CONTROLLER_ANALOG_LEFT_Y)) / 127;
+    double turn = (double)(joystick.get_analog(pros::E_CONTROLLER_ANALOG_RIGHT_X)) / 127;
     this->chassis->model().arcade(forward, turn);
 }
